+++ CakeShop/src/main/repository/database/OrderDbRepository.java
@@ -10,7 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class OrderDbRepository implements IRepository<Long, Order<Long>> {
+public class OrderDbRepository implements IRepository<Integer, Order<Integer>> {
     private final String url;
 
     public OrderDbRepository(String url) {
@@ -18,12 +18,12 @@ public class OrderDbRepository implements IRepository<Long, Order<Long>> {
     }
 
     @Override
-    public Long add(Order<Long> order) {
+    public Integer add(Order<Integer> order) {
         String sql = "INSERT INTO Orders (cakeId, customerName, quantity, status, createdAt) VALUES (?, ?, ?, ?, ?)";
         try (Connection conn = DriverManager.getConnection(url);
              PreparedStatement pstmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
 
-            pstmt.setLong(1, order.getCakeId());
+            pstmt.setInt(1, order.getCakeId());
             pstmt.setString(2, order.getCustomerName());
             pstmt.setInt(3, order.getQuantity());
             pstmt.setString(4, order.getStatus());
@@ -32,31 +32,32 @@ public class OrderDbRepository implements IRepository<Long, Order<Long>> {
             pstmt.executeUpdate();
             ResultSet rs = pstmt.getGeneratedKeys();
             if (rs.next()) {
-                long id = rs.getLong(1);
+                Integer id = rs.getInt(1);
                 order.setId(id);
                 return id;
             }
         } catch (SQLException e) {
             e.printStackTrace();
         }
+
         return null;
     }
 
     @Override
-    public Iterable<Order<Long>> findAll() {
-        List<Order<Long>> orders = new ArrayList<>();
+    public Iterable<Order<Integer>> findAll() {
+        List<Order<Integer>> orders = new ArrayList<>();
         String sql = "SELECT * FROM Orders";
         try (Connection conn = DriverManager.getConnection(url);
              Statement stmt = conn.createStatement();
              ResultSet rs = stmt.executeQuery(sql)) {
 
             while (rs.next()) {
-                Order<Long> order = new Order<>(
-                        rs.getLong("cakeId"),
+                Order<Integer> order = new Order<>(
+                        rs.getInt("cakeId"),
                         rs.getString("customerName"),
                         rs.getInt("quantity")
                 );
-                order.setId(rs.getLong("id"));
+                order.setId(rs.getInt("id"));
                 order.setStatus(rs.getString("status"));
                 order.setCreatedAt(LocalDateTime.parse(rs.getString("createdAt"), DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")));
                 orders.add(order);
@@ -68,20 +69,20 @@ public class OrderDbRepository implements IRepository<Long, Order<Long>> {
     }
 
     @Override
-    public Optional<Order<Long>> findById(Long id) {
+    public Optional<Order<Integer>> findById(Integer id) {
         String sql = "SELECT * FROM Orders WHERE id = ?";
         try (Connection conn = DriverManager.getConnection(url);
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
-            pstmt.setLong(1, id);
+            pstmt.setInt(1, id);
             ResultSet rs = pstmt.executeQuery();
             if (rs.next()) {
-                Order<Long> order = new Order<>(
-                        rs.getLong("cakeId"),
+                Order<Integer> order = new Order<>(
+                        rs.getInt("cakeId"),
                         rs.getString("customerName"),
                         rs.getInt("quantity")
                 );
-                order.setId(rs.getLong("id"));
+                order.setId(rs.getInt("id"));
                 order.setStatus(rs.getString("status"));
                 order.setCreatedAt(LocalDateTime.parse(rs.getString("createdAt"), DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")));
                 return Optional.of(order);
@@ -93,17 +94,17 @@ public class OrderDbRepository implements IRepository<Long, Order<Long>> {
     }
 
     @Override
-    public void modify(Order<Long> updatedOrder) {
+    public void modify(Order<Integer> updatedOrder) {
         String sql = "UPDATE Orders SET cakeId = ?, customerName = ?, quantity = ?, status = ?, createdAt = ? WHERE id = ?";
         try (Connection conn = DriverManager.getConnection(url);
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
-            pstmt.setLong(1, updatedOrder.getCakeId());
+            pstmt.setInt(1, updatedOrder.getCakeId());
             pstmt.setString(2, updatedOrder.getCustomerName());
             pstmt.setInt(3, updatedOrder.getQuantity());
             pstmt.setString(4, updatedOrder.getStatus());
             pstmt.setString(5, updatedOrder.getCreatedAt().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")));
-            pstmt.setLong(6, updatedOrder.getId());
+            pstmt.setInt(6, updatedOrder.getId());
 
             pstmt.executeUpdate();
         } catch (SQLException e) {
@@ -112,12 +113,12 @@ public class OrderDbRepository implements IRepository<Long, Order<Long>> {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Integer id) {
         String sql = "DELETE FROM Orders WHERE id = ?";
         try (Connection conn = DriverManager.getConnection(url);
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
-            pstmt.setLong(1, id);
+            pstmt.setInt(1, id);
             pstmt.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
