+++ CakeShop/src/main/repository/memory/InMemoryBirthdayCakeRepository.java
@@ -2,5 +2,5 @@ package main.repository.memory;
 
 import main.domain.BirthdayCake;
 
-public class InMemoryBirthdayCakeRepository extends InMemoryRepository<Long, BirthdayCake<Long>> {
+public class InMemoryBirthdayCakeRepository extends InMemoryRepository<Integer, BirthdayCake<Integer>> {
 }
