+++ CakeShop/src/main/repository/memory/InMemoryOrderRepository.java
@@ -2,5 +2,5 @@ package main.repository.memory;
 
 import main.domain.Order;
 
-public class InMemoryOrderRepository extends InMemoryRepository<Long, Order<Long>> {
+public class InMemoryOrderRepository extends InMemoryRepository<Integer, Order<Integer>> {
 }
